+++ tests/steps/field/test_field_remove.py
@@ -1,0 +1,25 @@
+from frictionless import Resource, transform, steps
+
+
+# General
+
+
+def test_step_field_remove():
+    source = Resource(path="data/transform.csv")
+    target = transform(
+        source,
+        steps=[
+            steps.field_remove(names=["id"]),
+        ],
+    )
+    assert target.schema == {
+        "fields": [
+            {"name": "name", "type": "string"},
+            {"name": "population", "type": "integer"},
+        ]
+    }
+    assert target.read_rows() == [
+        {"name": "germany", "population": 83},
+        {"name": "france", "population": 66},
+        {"name": "spain", "population": 47},
+    ]
